// src/charts/src/smithchart/model/interface.ts
export interface ISmithchartEventArgs {
    /** Defines the name of the event */
    name: string;
    /** Defines the event cancel status */
    cancel: boolean;
}

export interface ISmithchartPrintEventArgs extends ISmithchartEventArgs {
    htmlContent: Element;
}